requirements.txt
fastapi
uvicorn
requests
pandas
scikit-learn
xgboost
joblib
python-dotenv
nltk
beautifulsoup4
